requirements.txt
Flask
Pillow
streamlit
